fix: use maxType when building expression in expression tree elaborator
After the expression tree elaborator inserts coercions, it builds the expression itself. It was not storing the computed maxType in the return types of operators. This led to unexpected errors in examples such as
```lean
example (a : Nat) (b : Int) :
  a = id (a * b^2) := sorry
```
where it would say it could not synthesize an `HMul Int Int Nat` instance. The issue in this case is that `HPow` uses default instances, so while the expression tree elaborator decides that `a * b^2` should be referring to an `Int`, the actual elaborated type is temporarily a metavariable, so when the binrel elaborator is looking at both sides of the equality, it decides that `Nat` will work and coercions don't need to be inserted.

The fix is to have the expression tree elaborator elaborate all the operators using the computed `maxType` as the expected type.