feat: abstractMVars: do not loop on bad input
previously `abstractMVars` would go into a loop on bad input, in
particular if an mvar occurs in its own type. While this should never
happen in practice and is thus not a problem for our users, it can be
quote annoying when one runs into this problem while developing meta
code, as the stack overflow means you don’t get your `trace` messages
etc.

By keeping track of the mvars whose type we are processing we can break
the loop. The function is pure, so no `throwError`, but even a `panic!`
is going to be somewhat helpful.