fix: show argument name in implicit argument error (#4426)
When an implicit argument cannot be inferred, the error should show the
name of the argument.

Showing the argument name in the error message for an uninstantiated
metavariable was introduced in da33f498f539c4147e2254f05f8bd403257411c8,
but this implementation causes some argument names to get lost.