feat: more infrastructure for tactic documentation
Introduces the notion of a tactic alias, which is a syntax
specification that is really "the same as" an existing tactic, but
needs to be separate for technical reasons.

Additionally, tactics can be tagged, and extensibility is first-class.

This is a minimal proof of concept for evaluation - missing features
include:
 * Hovers on tactics don't take this new metadata into account
 * There's not yet a user interface for querying the data internally
 in Lean